harmonize setting output track numbers in writers
Track numbers that are not set are initialized just before creating
the files. The track number equals the last track number + 1
Track filenames in AS02 and Avid are created using an
initial / default track number. This means the number in the filename
could differ from the material package TrackNumber
Track numbers are now also settable in raw2bmx for the Avid and AS02 format
D10 AES-3 track numbers are checked so that there are no duplicates