as02/avid: calc duration using sample sequence for pcm tracks
This fixes problem where the clip duration, which is the minimum
track duration, is less than expected when using raw2bmx.

This problem is exposed when the clip rate does not equals the track
rate, i.e. when writing mono essence files in the avid and as02
writers.

This way makes the durations consistent when reading and writing.