reader: fix the read limit duration
The read limit duration includes every frame that is read, i.e
it includes precharge and rollout

(The material package and track duration do not include precharge
and rollout. The file track duration does not include rollout. The
container duration includes both precharge and rollout and is the
default read duration)