reader: apply limits at track level to precharge/rollout calculations
Added a GetAvailableReadLimits method to the readers to provide
the min and max extents of the sample data. These extents are used
to limit the amount of precharge or rollout that result from
not all the tracks. E.g. a mpeg track has 1 frame precharge but the
audio track has no precharge and no sample data and so the precharge
must be limited to 0 if limit_to_available is true.