bmxtranswrap: add --no-precharge --no-rollout options
If --no-precharge then adjust the start position and duration if
there is precharge. If --no-rollout then adjust the duration if
there is rollout.

Print info log message if start position or durations have been
adjusted. Print warn log message if the clip type does not support
precharge and/or rollout