avid: change image offset to end for alpha tracks
The Avid alpha tracks have align to the start of 8k blocks rather
than to the end.

The Avid alpha sample files didn't set the ImageEndOffset and so
this example if followed in the PictureMXFDescriptorHelper

Removed start offset padding in AvidTrack::WriteSamples which was wrong
for multiple samples. It is assumed that method is only called when
there is no padding required.