remove d10 padding bytes beyond max size
SMPTE ST386 specifies a maximum bit rate and therefore a maximum
D-10 frame size. If the frame size exceeds the maximum then the
assumption is that it includes padding bytes at the end of the frame
and these will now be removed in this commit

This removes the padding bytes that Avid adds to D-10 frames to
get 4k alignment.

Added --no-d10-depad option to raw2bmx and bmxtranswrap to disable
padding removal

Logging warning when padding removal disabled and the frame size exceeds
the maximum allowed size.

Logging info message if padding was added or removed