mxf reader: read frame metadata from essence container
Reading System Scheme 1 metadata: timecode array and crc-32 array.
The archive preservation project uses the timecode array to store
VITC and LTC and the crc-32 array stores crc-32 for each picture
and sound element.

Reading SDTI-CP system metadata pack: package rate and user
timecode. The package rate is needed to interpret the timecode and
the user timecode is used in XDCAM for LTC and could be used to
check for discontinuities.

Reading SDTC-CP package metadata pack: umid and essence marls. The
umid can be used to detect content that comes from different source
packages and the (Sony XDCAM) essence marks are useful to know when
the recording started, good shot marks etc.