mxf reader: use fsp edit rate rather than descriptor sample rate
Telestream Episode sample files (Company: OpenCube,
Product: MXFTk Advanced, Version: 2.4.1.20111221) had an incorrect
SampleRate property value in the file descriptors: 48000/1 when
it should have been 25/1. This update uses the top level file source
package's track edit rate instead.

Using the track edit rate is probably more likely to be correct
because it is assumed in the code to equal the index edit rate.
The sample rate (non-divisible, contiguously accessible units),
could be less than the edit rate depending on the whims of the
essence container specification author.

An exception is thrown if a track edit rates are different in the
sample internal file source package. A info message is logged is the
descriptor's sample rate differs from the fsp track edit rate.
Normally these are equal. A "not supported"  warning is logged if
the fsp track edit rate does not equal the material package track
edit rate.