apps/avid: add --import option and create only single import or tape package
An import source package is only created by explicitly providing
the filename or URI using the --import option.

An Avid clip may have a single import or tape source package

An absolute URL is filled in the network locator because Avid doesn't
accept relative URLs; the "Source File" bin property is only filled
when the URL is an absolute one.

Note that there are still a number of things missing to support
relinking based on import packages:
 - a timecode track is not created by default; this must change for
   sources that have a timecode track
 - the track ids of the import source package must match the track
   ids of the original (sourec?) package imported into Avid.
   Otherwise Avid will fail to find the track to relink with.
   Extracting this info from an input MXF file is possible. It is
   unclear what to do when the original is a quicktime file. Are the
   quicktime track numbers used or does Avid create its own set?