h264dump: add byte to NAL to workaround missing stop bit
Add a byte to the NAL unit bytes as a workaround for the missing
sequence parameter set stop bit in Avid Transfer Manager AVCI
files. The last couple of properties in the SPS are zero and the
last byte is wrongly assumed to be padding because of the missing
stop bit. This results in not be enough bits being available and
parsing fails