replace (v)snprintf with bmx_(v)snprintf
MS Visual C++ _(v)snprintf implementation differs from the POSIX
definition. The return value is different and a NULL terminator
is not written when the number of chars required >= size parameter.

The bmx_(v)snprintf implementation guarantees that the output
string is NULL terminated if string != NULL and size > 0. It
returns void.