build: add top src dir to include path
This fixes 'make distcheck' when not running in the top directory.

The CFLAGS has an include path to config.h and this means
'make distcheck' is OK if bmx_scm_version.h is in the same directory as
config.h. When running 'make distcheck' in some other directory there
is no include path to bmx_scm_version.h and 'make distcheck' fails

This issue was exposed by the 057bc09c because GIR_DIR no longer
points to a git repository directory and so an additional (there is
still the version from the distribution package) bmx_scm_version.h
is not generated.