d10: always write the maximum frame size
No longer matters what the input D-10 frame size is. The frame size
will be padded out to the maximum size for the given bit rate
(30/40/50Mbps) and frame rate. Input frames size exceeding the
maximum size, e.g. Avid OP-Atom, are assumed to contain padding and
that padding is checked and removed.