reader: support files with no index table (IndexSID=0)
Goes part of the way to supporting growing / incomplete files.

Supports frame wrapped essence and clip wrapped essence (with
known and fixed edit unit size) in single partition. Supports
'raw' XDCAM files (25Hz)

Changes include seeking to the start of the essence data
after opening the essence reader. This ensures that the file is
positioned correctly when reading in
MXFFileReader::ExtractInfoFromFirstFrame. This was not an issue
before because the essence reader adjusted the file position in
EssenceReader::ReadFrameWrappedSamples after getting the first
edit unit.