reader: handle re-read of last frame after failure
A read failure means the file position is assumed not to be at the
start of the content package. If the same frame is re-read then
seek back to the known frame start. The frame isn't indexed because
this only happens after a successful read and the frame size is
known