mpeg2lg index: handle B frames commence closed GOP case
Assume only backward prediction is used in the B frames at the
start of a closed GOP.

This was required for Adobe Media Encoder CS6, because without it
the encoder would just hang. It also brings it closer to the
RDD9 requirement for strict settings for index entry flags.

Also fixed the mBPictureCount type which should be UInt16