reader: remove descriptor sample rate = fsp edit rate check and info log
It appears that there are 2 camps in the MXF world, one which sets
the audio descriptor sample rate equal to the file source
package / video frame rate and the other sets it to the audio
sampling rate. See the last paragraph of ST382, section 7.2 and
section F.2 of ST377-1-2011 for contradictory statements.

The property is not used anyway and the check and info log have been
removed to reduce the noise.