reader: as02: handle invalid network locator urls
Avid Media Composer 6.5 doesn't escape the URLs for the AS-02 media
files in the MXF Network Locator URL String property. (It also
doesn't do it for the values in the manifest file). This results
in a failure to parse the value when it contains URI reserved
characters, e.g. the Avid sequence name that is used as the prefix
for the version and media filenames contains spaces.

This updates follows the recomendations of ST377-1-2011, section
B.4.4 "handling invalid or unknown URLs and URIs". A URL String
that cannot be parsed and doesn't start with 'file://' is assumed
to be a file path.