reader: workaround non-zero IndexSID but no index table in file
Issue reported by Stefan Schilk on SF forum. Reported that the files
originated from an Admira partial restore system.

ST377-1-2011, section 11.5.1 specifies that an index table shall be
present if BodySID and IndexSID are non-zero.

The IndexTableHelper is only marked as complete if there are index
tables segments present in the complete file. An exception is made
if the essence container size is 0 in a complete file.