restructure as11 writing classes
Reduces code duplication in the AS11Clip/AS11Track classes wrt
the ClipWriter/ClipWriterTrack classes. Added a AS11WriterHelper
class that sits alongside the ClipWriter; it doesn't contain or
wrap the writer, but instead provides methods to add metadata to
a clip.

Removed the AS11 clip writer type enumerations and replaced it with an
apps clip sub type enumeration to narrow the clip type to AS-11.