op1a/as11op1a: don't reset input duration for single pass failure
The mInputDuration is no longer reset to -1 when a complete header
partition cannot be achieved when the single pass write flavour
is enabled. The mInputDuration variable is only accessed when
mSupportCompleteSinglePass is true.

The update allows an AS-11 segmentation track to be completed using
the input duration.

Also improved warning messages when the header partition cannot
be completed when single pass flavour is enabled.