reader: improve essence wrapping type detection
First check for known wrapping type using the essence container
label. If it is unknown, then make an initial guess based on the
operational pattern.

If the wrapping type has been guessed, then look at the first frame
to see whether the wrapping type can be detected. Extract the index table
segments before the essence container layout. If the first edit unit's
size is known from the index table, then compare it with the first
essence element KLV length. If it is >= L then it must be clip-wrapped,
else frame-wrapped.