mxf2raw: use getDescriptorLight to avoid dereference failure
Some files have unknown source package descriptors and these
should not result in fatal errors, particularly because they are
often physical source packages.

Also, don't assume it is a physical source package when the descriptor
cannot be dereferenced. Only include physical source package info
if the descriptor can be dereferenced and it is a subclass of
PhysicalDescriptor.