mxf2raw: treat duration/position differently and duration < 0
A property of duration type has -1 as its distinguished value to
indicate the duration is unknown. In XML this will result in an
empty hh:mm:ss:ff string and the count attribute equals -1. In text
the 'unknown' string will be used in place of hh:mm:ss:ff.

A property of position type can have negative values. The hh:mm:ss:ff
for a negative position in XML and text is represented as prefixing
the string with a '-'.