set last byte of vbi/anc ess key and track num to 01h
Tables in ST 436 show essence element key and track numbers with
the last byte set to 01h. The generic container spec. ST 379-1
states that the last byte can range from 00h-7fh. It is unclear
whether the value in ST 436 was intended to restrict ST 379-1.

The last byte is now set to 01h to avoid possible compatibility
issues with readers that expect that value.