op1a: change default to 0 mp track number
The material package track number is only used (set to a non-zero value)
when an AS-11 Core Framework is provided to bmxtranswrap/raw2bmx.
This track number is needed to interpret the audio track layout property
in that framework.

The --zero-mp-track-num option was replaced with --mp-track-num. The option
can be used to enable setting material package track numbers.

The AS-02 clip type is unchanged and will set the track number.