Skip MXF KL found in D-10 essence
Quicktime files sometimes include the MXF KL for D-10 in the MPEG
essence data. These additional bytes need to be removed when
converting to MXF via raw2bmx. This change modifies the raw essence reader
used by raw2bmx (and bmxparse) to check the start to see
if there is a D-10 KL. If there is then it uses the
KLVEssenceSource to surface the V only.