Support transwrap to wave if less samples than transfer rate requires
If an input file contains clip wrapped PCM data and the number of
samples is less than a multiple of the samples per frame then this
commit will ensure the partial frame can be written to a WAVE file.

E.g. the input has 960 audio samples at 48kHz. bmxtranswrap
defaults to a 25 Hz frame rate and the first frame would be incomplete
with only 960 of 1920 samples available. If the output is a WAVE file
then bmxtranswrap will accept the partial frame containing 960 samples.

Also added tests for the existing capability for adding pcm padding
data when transwrapping Avid MXF to OP1a where the Avid has video
pre-charge and there is no audio data available in the pre-charge range.