Ensure git checkout uses LF for test text files
git on Windows (2.38.1.windows.1) was found to have 'core.autocrlf' set to 'true'
(in the global gitconfig), which resulted in files detected as text to be be
checked out (e.g. as part of the repo clone) with CRLF line endings. That then
broke the tests that rely on checksums and reference text files because the input
data had changed.

The use of .gitattributes and "text eol=lf" attributes did nothing to change
that for the existing test files! Confusingly it did work for new
files checked into the repo?! 'git check-attr -a' and 'git ls-files --eol'
did not show any diference between the existing and new files.

This commit does the "proper thing", which is use .gitattributes
to indicate files that should be checked out with LF line endings.
To workaround git ignoring these attributes, a '.bin' suffix is added to
the text filenames to stop git detecting it as text and using CRLF anyway
on checkout.