Use HTTP GET request to get file size if HEAD request fails
This supports pre-signed URLs that restrict the request method. A HEAD
request would fail in that case.

Knowing the file size allows bmx to parse the RIP to get an index of the
partitions. The fallback is to seek to the footer partition and then
work back through the partitions, but that doesn't work in some files which
don't set the footer partition offset.