Workaround HTTP2 connection error
It was found that the google storage api would cause curl to return
a CURLE_HTTP2 error after around an hour. The workaround implemented here
is to close the connection and open a new one. The issue appears similar
to https://github.com/curl/curl/issues/5389 and
https://github.com/curl/curl/pull/5643. The fix
https://github.com/curl/curl/commit/ef86daf4d39e99b227d42bb712000c9adfdbdf76
didn't solve the issue (version 7.74.0).