rdd9: raise partition duration limit check to allow matching spec values
RDD 9 2013 (and 2009) specify an approximate 10 second partition
duration limit, not a hard maximum.

This commit will result in a partition duration matching the values in
section B.6 if a fixed GOP size of 12 or 15 frames is used. If a
different GOP structure is used then the partition may be larger (up to
15 frames) than the value in the spec (because a partition needs to
start with a key frame), but that's fine as the 10 second is not a hard
limit.