app: add --disable-indexing-file option for faster file access
Using this option will stop the reader indexing the locations of essence.
That is useful if there are many partitions for example and it's more
efficient to not index and instead treat the file as incomplete. It may
not be efficient if essence is read and the --start option is used with a
position not near the start.

The footer is still searched for and used to extract header metadata
(if available). This ensures that the most up-to-date metadata is used.
The footer is found either using the RIP or the footerPartition property
in the header partition pack.