Don't include "manufacturer" attribute in RDD 6 XML output
mxf2raw will no longer include it in the RDD 6 XML it outputs.
bmxtranswrap and raw2bmx will continue to ignore it and the
XML Schema still allows it to be present.

The "manufacturer" string attribute was included for known integer
values, but it is not part in the RDD 6 serial bitstream.

This attribute may be removed in future.