Add guards for NULL data ptr pass to fread/fwrite
Only required in places where the origin of data is unknown and
therefore could be provided by a library user.

This avoids a sanitizer warning about the argument being flagged as
non-NULL when there is a test that passes NULL. GCC appears to be
fine with it but other compilers may not as the behaviour is undefined.