Add support for separate field coded AVC
Added ParseFrameSize2() and ParseFrameInfo2() essence parser methods
that are given / return a ParsedFrameSize object that holds frame parse
state and supports separate fields. These methods are called rather than
the old (version 1) methods and if a parser does not implement them
(which is all but the AVC parser) then the new method calls the old
method.

The previous AVCEssenceParser was renamed to
AVCCodedPictureEssenceParser and it is limited to parsing a coded
picture (frame coded or field coded).

The AVCEssenceParser was changed to support frames or separate field pairs.

The AVCWriteHelper was updated to handle the fields (access units /
codec pictures) when calculating the index and AVC sub-descriptor properties.